futures/requirements_futures.txt
# ============================================================
# futures/ 涓撶敤渚濊禆
# 鍚堣澹版槑锛氫笉寰楀寘鍚换浣曞姞瀵嗚揣甯佺浉鍏冲寘锛坈cxt, web3, binance 绛夛級
# AppID: simnow_client_test
# ============================================================

# CTP 鎺ュ彛
vnpy_ctp>=3.7.0

# 鏍稿績寮傛
aiohttp>=3.9.0
aiosqlite>=0.19.0

# 鏁版嵁缁撴瀯
pydantic>=2.5.0
pyyaml>=6.0

# 鎶€鏈垎鏋?
pandas>=2.1.0
numpy>=1.26.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0

# 鏃ュ織
loguru>=0.7.0
rich>=13.7.0

# 鐜鍙橀噺
python-dotenv>=1.0.0

# 娴嬭瘯
pytest>=7.4.0
pytest-asyncio>=0.23.0

